Added rudimentary AI
The AI is a basic minimax algorithm. It performs no optimizations at
all. Therefore, it is excruciatingly slow at depths of 3 or over.

It also acts in a very unnatural manner: it picks the first spot from
the top left it can go into for a long time. That's because it cannot
identify the potentially huge squares that it can create because the
depth is too low.

For such a poor AI it might be possible that the introduction of a
randomizer will actually IMPROVE it.

This is left to be explored in subsequent versions.

Hopefully I will also somehow optimize the score calculation algorithm
to make it run in linear time.